refactor: remove habit metrics in bulk (#370)

+++ src/services/metrics.service.ts
@@ -46,11 +46,15 @@ export const patchHabitMetric = async (
   return deepCamelcaseKeys<HabitMetric>(data);
 };
 
-export const destroyHabitMetric = async (id: string) => {
+export const destroyHabitMetrics = async (ids: string[]) => {
+  if (ids.length === 0) {
+    return;
+  }
+
   const { error } = await supabaseClient
     .from('habit_metrics')
     .delete()
-    .eq('id', id);
+    .in('id', ids);
 
   if (error) {
     throw new Error(error.message);
